✨ Payable and Assignor Data Components
